+++ lib/coinbase.rb
@@ -21,6 +21,7 @@ require_relative 'coinbase/transaction'
 require_relative 'coinbase/user'
 require_relative 'coinbase/wallet'
 require_relative 'coinbase/server_signer'
+require_relative 'coinbase/sponsored_send'
 require_relative 'coinbase/staking_operation'
 require_relative 'coinbase/staking_reward'
 require_relative 'coinbase/validator'
@@ -130,6 +131,17 @@ module Coinbase
     raise Coinbase::APIError.from_error(e), cause: nil
   rescue StandardError => e
     raise e
+  end
+
+  # Returns a pretty-printed object string that contains the object's class name and
+  # the details of the object, filtering out nil values.
+  # @param klass [Class] the class of the object
+  # @param details [Hash] the details of the object
+  # @return [String] the pretty-printed object string
+  def self.pretty_print_object(klass, **details)
+    filtered_details = details.filter { |_, v| !v.nil? }.map { |k, v| "#{k}: '#{v}'" }
+
+    "#{klass}{#{filtered_details.join(', ')}}"
   end
 
   # Returns whether to use a server signer to manage private keys.