+++ spec/support/shared_examples/address_staking.rb
@@ -9,7 +9,7 @@ shared_context 'with mocked staking_balances' do
     instance_double(
       Coinbase::Client::StakingContext,
       context: instance_double(
-        Coinbase::Client::PartialEthStakingContext,
+        Coinbase::Client::StakingContextContext,
         stakeable_balance: build(:balance_model, whole_amount: stake_balance),
         unstakeable_balance: build(:balance_model, whole_amount: unstake_balance),
         claimable_balance: build(:balance_model, whole_amount: claim_stake_balance)