+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `unsigned_payload`, `signed_payload`, `status`, and `transaction_hash` in-line fields on `Transfer` are deprecated in
   favor of those on `Transaction`
 
+### Added
+
+- Add network constants for Base Mainnet, Ethereum Holesky, and Ethereum Mainnet
+
 ## [0.0.13] - 2024-07-30
 
 ### Added